+++ yii2-adapter/legacy/widgets/MissingWidget.php
@@ -6,15 +6,15 @@
  * @license https://craftcms.github.io/license/
  */
 
+declare(strict_types=1);
+
 namespace craft\widgets;
 
-/** @phpstan-ignore-next-line */
-if (false) {
-    /**
-     * @since 3.0.0
-     * @deprecated in 6.0.0. Use `\CraftCms\Cms\Dashboard\Widgets\MissingWidget` instead.
-     */
-    class MissingWidget extends \CraftCms\Cms\Dashboard\Widgets\MissingWidget
-    {
-    }
+/**
+ * @since 3.0.0
+ * @deprecated in 6.0.0. Use `\CraftCms\Cms\Dashboard\Widgets\MissingWidget` instead.
+ */
+class MissingWidget extends \CraftCms\Cms\Dashboard\Widgets\MissingWidget
+{
+    use BuiltInWidgetTrait;
 }